Section headers, EventCategories now enum class, ClipBoard -> Clipboard, event dispatcher optimization, more includes in Flare.h.

+++ Flare/src/Flare/Events/ApplicationEvent.h
@@ -6,21 +6,21 @@ namespace Flare::Events
 {
    class FLARE_API ApplicationEvent : public Event
    {
-      /*****   CLASS   C-TOR D-TOR  *****/
+      /*****  C-TOR D-TOR  *****/
       protected:
       ApplicationEvent() {}
 
       /*****   GETTERS   *****/
-      FLARE_EVENT_CATEGORY( EventCategory::Application );
+      FLARE_EVENT_CATEGORY( static_cast<uint32_t>(EventCategory::Application) );
    };
 
    class FLARE_API AppTickEvent : public ApplicationEvent
    {
-      /*****   CLASS   C-TOR D-TOR  *****/
+      /*****  C-TOR D-TOR  *****/
       public:
       AppTickEvent() : ApplicationEvent() {}
 
-      /*****   CLASS   FUNCTIONS    *****/
+      /*****   FUNCTIONS   *****/
       std::string ToString() const
       {
          std::stringstream x_SS;
@@ -34,11 +34,11 @@ namespace Flare::Events
 
    class FLARE_API AppBeforeUpdateEvent : public ApplicationEvent
    {
-      /*****   CLASS   C-TOR D-TOR  *****/
+      /*****  C-TOR D-TOR  *****/
       public:
       AppBeforeUpdateEvent() : ApplicationEvent() {}
 
-      /*****   CLASS   FUNCTIONS    *****/
+      /*****   FUNCTIONS   *****/
       std::string ToString() const
       {
          std::stringstream x_SS;
@@ -52,11 +52,11 @@ namespace Flare::Events
 
    class FLARE_API AppUpdateEvent : public ApplicationEvent
    {
-      /*****   CLASS   C-TOR D-TOR  *****/
+      /*****  C-TOR D-TOR  *****/
       public:
       AppUpdateEvent() : ApplicationEvent() {}
 
-      /*****   CLASS   FUNCTIONS    *****/
+      /*****   FUNCTIONS   *****/
       std::string ToString() const
       {
          std::stringstream x_SS;
@@ -70,11 +70,11 @@ namespace Flare::Events
 
    class FLARE_API AppAfterUpdateEvent : public ApplicationEvent
    {
-      /*****   CLASS   C-TOR D-TOR  *****/
+      /*****  C-TOR D-TOR  *****/
       public:
       AppAfterUpdateEvent() : ApplicationEvent() {}
 
-      /*****   CLASS   FUNCTIONS    *****/
+      /*****   FUNCTIONS   *****/
       std::string ToString() const
       {
          std::stringstream x_SS;
@@ -88,11 +88,11 @@ namespace Flare::Events
 
    class FLARE_API AppBeforeRenderEvent : public ApplicationEvent
    {
-      /*****   CLASS   C-TOR D-TOR  *****/
+      /*****  C-TOR D-TOR  *****/
       public:
       AppBeforeRenderEvent() : ApplicationEvent() {}
 
-      /*****   CLASS   FUNCTIONS    *****/
+      /*****   FUNCTIONS   *****/
       std::string ToString() const
       {
          std::stringstream x_SS;
@@ -106,11 +106,11 @@ namespace Flare::Events
 
    class FLARE_API AppRenderEvent : public ApplicationEvent
    {
-      /*****   CLASS   C-TOR D-TOR  *****/
+      /*****  C-TOR D-TOR  *****/
       public:
       AppRenderEvent() : ApplicationEvent() {}
 
-      /*****   CLASS   FUNCTIONS    *****/
+      /*****   FUNCTIONS   *****/
       std::string ToString() const
       {
          std::stringstream x_SS;
@@ -124,11 +124,11 @@ namespace Flare::Events
 
    class FLARE_API AppAfterRenderEvent : public ApplicationEvent
    {
-      /*****   CLASS   C-TOR D-TOR  *****/
+      /*****  C-TOR D-TOR  *****/
       public:
       AppAfterRenderEvent() : ApplicationEvent() {}
 
-      /*****   CLASS   FUNCTIONS    *****/
+      /*****   FUNCTIONS   *****/
       std::string ToString() const
       {
          std::stringstream x_SS;
