+++ Flare/src/Flare/Events/KeyEvent.h
@@ -6,11 +6,11 @@ namespace Flare::Events
 {
    class FLARE_API KeyEvent : public Event
    {
-      /*****   CLASS   VARIABLES    *****/
+      /*****   VARIABLES   *****/
       protected:
       int m_KeyCode;
 
-      /*****   CLASS   C-TOR D-TOR  *****/
+      /*****  C-TOR D-TOR  *****/
       protected:
       KeyEvent( int a_KeyCode ) :
          m_KeyCode( a_KeyCode )
@@ -19,23 +19,23 @@ namespace Flare::Events
       /*****   GETTERS   *****/
       public:
       inline int GetKeyCode() const { return m_KeyCode; }
-      FLARE_EVENT_CATEGORY( EventCategory::Keyboard | EventCategory::Input );
+      FLARE_EVENT_CATEGORY( static_cast<uint32_t>( EventCategory::Keyboard ) | static_cast<uint32_t>( EventCategory::Input ) );
    };
 
    class FLARE_API KeyPressedEvent : public KeyEvent
    {
-      /*****   CLASS   VARIABLES    *****/
+      /*****   VARIABLES   *****/
       private:
       int m_RepeatCount;
 
-      /*****   CLASS   C-TOR D-TOR  *****/
+      /*****  C-TOR D-TOR  *****/
       public:
       KeyPressedEvent( int a_KeyCode, int a_RepeatCount ) :
          KeyEvent( a_KeyCode ),
          m_RepeatCount( a_RepeatCount )
       {}
 
-      /*****   CLASS   FUNCTIONS    *****/
+      /*****   FUNCTIONS   *****/
       public:
       std::string ToString() const override
       {
@@ -52,13 +52,13 @@ namespace Flare::Events
 
    class FLARE_API KeyReleasedEvent : public KeyEvent
    {
-      /*****   CLASS   C-TOR D-TOR  *****/
+      /*****  C-TOR D-TOR  *****/
       public:
-      KeyReleasedEvent( int a_KeyCode) :
+      KeyReleasedEvent( int a_KeyCode ) :
          KeyEvent( a_KeyCode )
       {}
 
-      /*****   CLASS   FUNCTIONS    *****/
+      /*****   FUNCTIONS   *****/
       public:
       std::string ToString() const override
       {
@@ -74,13 +74,13 @@ namespace Flare::Events
 
    class FLARE_API KeyTypedEvent : public KeyEvent
    {
-      /*****   CLASS   C-TOR D-TOR  *****/
+      /*****  C-TOR D-TOR  *****/
       public:
       KeyTypedEvent( int a_KeyCode ) :
          KeyEvent( a_KeyCode )
       {}
 
-      /*****   CLASS   FUNCTIONS    *****/
+      /*****   FUNCTIONS   *****/
       public:
       std::string ToString() const override
       {
